Move plotting and distance functions to utils modules
