+++ src/data_quality_utils/tree_finder/__init__.py
@@ -1,3 +1,4 @@
+from .plotting_utils import show_image, show_stats
 from .tree_finder import TreeFinder
 
-__all__ = ["TreeFindex"]
+__all__ = ["TreeFinder", "show_image", "show_stats"]
