+++ src/data_quality_utils/tree_finder/tree_finder.py
@@ -12,9 +12,10 @@ import cv2
 import kagglehub
 import numpy as np
 from deepforest import main
-from geopy.distance import geodesic
 from matplotlib import pyplot as plt
-from matplotlib.figure import Figure
+
+from .distance_utils import calculate_distances
+from .plotting_utils import generate_image
 
 # silence some deepforest warnings
 warnings.filterwarnings(
@@ -26,9 +27,7 @@ warnings.filterwarnings(
 )
 
 logger = logging.getLogger(__name__)
-TILE_SIZE = 256
 MODEL_INFERENCE = {"patch_size": 5000, "patch_overlap": 0.9, "iou_threshold": 1}
-COLOR = "#00625E"
 
 
 @contextmanager
@@ -87,77 +86,6 @@ class TreeFinder:
         )
         return pred_boxes
 
-    def _calculate_distances(
-        self,
-        lat: float,
-        lon: float,
-        img_width: int,
-        img_height: int,
-        pred_boxes: list[dict[str, float]],
-        convert_coords: bool,
-        bbox: tuple[float, float, float, float] | None = None,
-        zoom: int | None = None,
-        scale: float | None = None,
-    ) -> list[tuple[tuple[float, float], float, dict[str, float]]]:
-        """Calculate distances from predicted tree boxes to a given geographic location."""
-        distances = []
-        for box in pred_boxes:
-            x_center = (box["xmin"] + box["xmax"]) / 2
-            y_center = (box["ymin"] + box["ymax"]) / 2
-
-            if convert_coords:
-                center_px, center_py = TreeFinder.epsg3857_to_pixel(
-                    lat, lon, zoom, scale
-                )
-                abs_px = center_px - (img_width / 2) + x_center
-                abs_py = center_py - (img_height / 2) + y_center
-                pred_lat, pred_lon = TreeFinder.pixel_to_epsg3857(
-                    abs_px, abs_py, zoom, scale
-                )
-            else:
-                pred_lon, pred_lat = TreeFinder.pixel_to_epsg4326(
-                    x_center, y_center, img_width, img_height, bbox
-                )
-
-            box_dist = geodesic((pred_lat, pred_lon), (lat, lon)).meters
-            distances.append(((pred_lat, pred_lon), box_dist, box))
-        return distances
-
-    def _generate_boxes(
-        self, image: np.ndarray, pred_boxes: list[dict[str, float]] | dict[str, float]
-    ) -> None:
-        """Draw bounding boxes on the image."""
-        thickness = max(1, image.shape[0] // 500)
-        if isinstance(pred_boxes, dict):
-            pred_boxes = [pred_boxes]
-        for box in pred_boxes:
-            cv2.rectangle(
-                image,
-                (int(box["xmin"]), int(box["ymin"])),
-                (int(box["xmax"]), int(box["ymax"])),
-                (255, 165, 0),
-                thickness=thickness,
-                lineType=cv2.LINE_AA,
-            )
-
-    def _generate_point(self, image: np.ndarray, point: tuple[int, int]) -> None:
-        """Draw a point on the image."""
-        radius = max(5, (image.shape[0] // 500) * 5)
-        cv2.circle(image, point, radius=radius, color=(0, 0, 255), thickness=-1)
-
-    def _generate_image(
-        self,
-        image: np.ndarray,
-        pred_boxes: list[dict[str, float]] | dict[str, float],
-        point: tuple[int, int] | None = None,
-    ) -> np.ndarray:
-        """Overlay bounding boxes and optional point on the image."""
-        image_copy = image.copy()
-        self._generate_boxes(image_copy, pred_boxes)
-        if point:
-            self._generate_point(image_copy, point)
-        return image_copy
-
     def _get_image_metadata(self, filename: str) -> dict | None:
         """Load and return metadata for an image."""
         metadada_filename = filename.replace(".png", ".json")
@@ -175,7 +103,7 @@ class TreeFinder:
         """
         image = cv2.imread(filename)
         pred_boxes = self._predict_boxes(image)
-        result_image = self._generate_image(image, pred_boxes)
+        result_image = generate_image(image, pred_boxes)
         return result_image
 
     def find_closest_tree(
@@ -206,7 +134,7 @@ class TreeFinder:
         scale = metadata.get("scale", 1)
         img_size = image.shape[:2]
 
-        distances = self._calculate_distances(
+        distances = calculate_distances(
             lat,
             lon,
             img_size[0],
@@ -228,45 +156,8 @@ class TreeFinder:
         else:
             x_pixel, y_pixel = TreeFinder.epsg4326_to_pixel(lat, lon, bbox, img_size)
 
-        result_image = self._generate_image(image_copy, best_box, (x_pixel, y_pixel))
+        result_image = generate_image(image_copy, best_box, (x_pixel, y_pixel))
         return best_dist, flagged, result_image
-
-    def _show_single(self, image: np.ndarray, image_title: str) -> Figure:
-        """Display a single image"""
-        fig, ax = plt.subplots(1, 1, figsize=(5, 5))
-        ax.imshow(image)
-        ax.set_title(image_title)
-        ax.axis("off")
-        return fig
-
-    def _show_multiple(
-        self, image_list: list[np.ndarray], image_title_list: list[str]
-    ) -> Figure:
-        """Display multiple images side by side."""
-        num_images = len(image_list)
-        fig, axes = plt.subplots(1, num_images, figsize=(5 * num_images, 5))
-        for ax, image, title in zip(axes, image_list, image_title_list):
-            ax.imshow(image)
-            ax.set_title(title)
-            ax.axis("off")
-        return fig
-
-    def show_image(
-        self,
-        images: np.ndarray | list[np.ndarray],
-        image_titles: str | list[str] | None = None,
-        save: bool = False,
-        save_path: str | None = None,
-    ) -> None:
-        """Show one or more images."""
-        if isinstance(images, list):
-            fig = self._show_multiple(images, image_titles)
-        else:
-            fig = self._show_single(images, image_titles)
-        if save:
-            os.makedirs(os.path.dirname(save_path), exist_ok=True)
-            fig.savefig(save_path, dpi=300, bbox_inches="tight")
-        plt.show()
 
     def get_stats(self, paths: list[str], types: list[str]) -> dict[str, list[float]]:
         """Compute distance statistics for tree predictions across multiple image sets."""
@@ -291,7 +182,7 @@ class TreeFinder:
                 scale = metadata.get("scale", 1)
                 img_size = image.shape[:2]
 
-                distances = self._calculate_distances(
+                distances = calculate_distances(
                     lat,
                     lon,
                     img_size[0],
@@ -309,85 +200,3 @@ class TreeFinder:
 
             stats[type] = type_stats
         return stats
-
-    @staticmethod
-    def show_stats(stats_dict: dict[str, list[float]]) -> None:
-        """Plot histograms of tree detection distances for each image type."""
-        max_distance = math.ceil(
-            max(max(type_stats) for type_stats in stats_dict.values())
-        )
-        num_stats = len(stats_dict)
-
-        bins = list(range(0, max_distance + 1))
-
-        fig, axes = plt.subplots(
-            num_stats, 1, sharex=True, figsize=(6, 6), height_ratios=[1, 1]
-        )
-        fig.subplots_adjust(hspace=0.1)
-
-        for ax, type, type_stats in zip(axes, stats_dict.keys(), stats_dict.values()):
-            ax.hist(type_stats, bins=bins, color=COLOR, label=f"{type} image")
-            ax.set_title(f"{type} error distribution")
-            ax.set_ylabel("Count")
-            # ax.legend(loc="best")
-
-        axes[-1].set_xlabel("Distance from true point (meters)")
-        axes[-1].set_xticks(bins)
-
-        plt.tight_layout()
-        plt.savefig("data/distance_histograms.png", dpi=300, bbox_inches="tight")
-        plt.show()
-
-    @staticmethod
-    def pixel_to_epsg4326(
-        x: float,
-        y: float,
-        img_width: int,
-        img_height: int,
-        bbox: tuple[float, float, float, float],
-    ) -> tuple[float, float]:
-        """Convert image pixel coordinates to EPSG:4326 (lat/lon)."""
-        xmin, ymin, xmax, ymax = bbox
-        lon = xmin + (x / img_width) * (xmax - xmin)
-        lat = ymax - (y / img_height) * (ymax - ymin)
-        return lon, lat
-
-    @staticmethod
-    def epsg4326_to_pixel(
-        lat: float,
-        lon: float,
-        bbox: tuple[float, float, float, float],
-        img_size: tuple[int, int],
-    ) -> tuple[int, int]:
-        """Convert EPSG:4326 (lat/lon) to pixel coordinates in an image."""
-        xmin, ymin, xmax, ymax = bbox
-        width, height = img_size
-        x_frac = (lon - xmin) / (xmax - xmin)
-        y_frac = 1 - (lat - ymin) / (ymax - ymin)
-        x_pixel = int(x_frac * width)
-        y_pixel = int(y_frac * height)
-        return x_pixel, y_pixel
-
-    @staticmethod
-    def epsg3857_to_pixel(
-        lat: float, lon: float, zoom: int, scale: float
-    ) -> tuple[float, float]:
-        """Convert EPSG:3857 coordinates to pixel coordinates."""
-        siny = math.sin(math.radians(lat))
-        siny = min(max(siny, -0.9999), 0.9999)
-
-        map_size = TILE_SIZE * (2**zoom) * scale
-        x = (lon + 180) / 360 * map_size
-        y = (0.5 - math.log((1 + siny) / (1 - siny)) / (4 * math.pi)) * map_size
-        return x, y
-
-    @staticmethod
-    def pixel_to_epsg3857(
-        x: float, y: float, zoom: int, scale: float
-    ) -> tuple[float, float]:
-        """Convert pixel coordinates to EPSG:3857 (lat/lon)."""
-        map_size = TILE_SIZE * (2**zoom) * scale
-        lon = x / map_size * 360.0 - 180.0
-        n = math.pi - 2.0 * math.pi * y / map_size
-        lat = math.degrees(math.atan(math.sinh(n)))
-        return lat, lon
